Feature/#24

+++ meter-tracker-web/src/main/java/com/asuprun/metertracker/web/service/IndicationService.java
@@ -44,20 +44,24 @@ public class IndicationService {
     private IndicationImageProcessor extractor;
     private DigitRecognizer recognizer;
 
-    @Autowired
     private IndicationRepository indicationRepository;
-
-    @Autowired
     private DigitRepository digitRepository;
-
-    @Autowired
     private ResourceRepository resourceRepository;
-
-    @Autowired
     private ResourceBindingRepository resourceBindingRepository;
+    private MeterRepository meterRepository;
 
     @Autowired
-    private MeterRepository meterRepository;
+    public IndicationService(IndicationRepository indicationRepository,
+                             DigitRepository digitRepository,
+                             ResourceRepository resourceRepository,
+                             ResourceBindingRepository resourceBindingRepository,
+                             MeterRepository meterRepository) {
+        this.indicationRepository = indicationRepository;
+        this.digitRepository = digitRepository;
+        this.resourceRepository = resourceRepository;
+        this.resourceBindingRepository = resourceBindingRepository;
+        this.meterRepository = meterRepository;
+    }
 
     @PostConstruct
     private void init() {
@@ -131,7 +135,7 @@ public class IndicationService {
     }
 
     public Optional<Indication> findById(long id) {
-        return Optional.ofNullable(indicationRepository.findOne(id));
+        return indicationRepository.findById(id);
     }
 
     /**
@@ -157,10 +161,10 @@ public class IndicationService {
     }
 
     public void delete(long id) {
-        findById(id).orElseThrow(() -> {
+        if (!indicationRepository.exists(id)) {
             logger.debug("Cannot delete not existing indication with id: {}", id);
-            return new NoSuchElementException("No such indication");
-        });
+            throw new NoSuchElementException("No such indication");
+        }
         indicationRepository.delete(id);
     }
 
@@ -188,21 +192,30 @@ public class IndicationService {
                     indication.getMeter().getCapacity(), digits.size());
             throw new IllegalArgumentException("Incorrect number of digits");
         }
-        digits.stream().filter(d -> d.getImage() != null).map(digitRepository::save).collect(toList());
-        String[] value = splitValue(digits.stream().map(Digit::getValue).collect(joining()), indication.getMeter());
+        digits.stream().filter(d -> d.getImage() != null).forEach(digitRepository::save);
 
-        indication.setValue(Double.parseDouble(String.join(".", value[0], value[1])));
+        double value = parseValue(digits.stream().map(Digit::getValue).collect(joining()), indication.getMeter());
+        indication.setValue(value);
+        indication.setConsumption(calculateConsumption(indication.getMeter().getId(), indication.getCreated(), value));
         indicationRepository.save(indication);
 
         trainRecognizer();
     }
 
-    private String[] splitValue(String value, Meter meter) {
+    private int calculateConsumption(long meterId, Date created, double value) {
+        indicationRepository.findRecognizedAfter(meterId, created).ifPresent(i -> {
+            i.setConsumption((int) Math.floor(i.getValue()) - ((int) value));
+            indicationRepository.save(i);
+        });
+
+        return indicationRepository.findRecognizedBefore(meterId, created)
+                .map(i -> ((int) value) - (int) Math.floor(i.getValue()))
+                .orElse(0);
+    }
+
+    private double parseValue(String value, Meter meter) {
         int splitIndex = meter.getCapacity() - meter.getMinorDigits();
-        return new String[]{
-                value.substring(0, splitIndex),
-                value.substring(splitIndex)
-        };
+        return Double.parseDouble(String.join(".", value.substring(0, splitIndex), value.substring(splitIndex)));
     }
 
     private void trainRecognizer() {
@@ -222,8 +235,8 @@ public class IndicationService {
             return DatatypeConverter.printHexBinary(digest.digest());
         } catch (NoSuchAlgorithmException e) {
             logger.warn("Cannot calculate image hash.", e);
+            throw new RuntimeException("Exception during image hash calculation", e);
         }
-        return null;
     }
 
 }
